requirements.txt
# ============================================
# PLAF - Prescriptive Learning Analytics Framework
# Requirements File
# ============================================

# -------------------- Data Processing --------------------
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# -------------------- Machine Learning --------------------
scikit-learn==1.3.2
imbalanced-learn==0.11.0  # For handling imbalanced datasets

# Vector Search
faiss-cpu==1.7.4  # For RAG similarity search

# ML Models
catboost==1.2.2
xgboost==2.0.3
lightgbm==4.1.0

# -------------------- eXplainable AI (XAI) --------------------
shap==0.44.0
anchor-exp==0.0.2.0
dice-ml==0.11

# -------------------- LLM Integration --------------------
google-generativeai==0.3.2  # Gemini API
python-dotenv==1.0.0  # For managing API keys

# Optional: LangChain (nếu muốn dùng advanced prompt chains)
# langchain==0.1.0
# langchain-google-genai==0.0.6

# -------------------- Visualization --------------------
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# -------------------- Dashboard --------------------
streamlit==1.29.0
# Hoặc thay bằng Gradio nếu thích:
# gradio==4.10.0

# -------------------- Utilities --------------------
joblib==1.3.2  # Model serialization
pyyaml==6.0.1  # Config files
tqdm==4.66.1  # Progress bars
tabulate==0.9.0  # Pretty tables in CLI

# -------------------- Jupyter Notebooks --------------------
jupyter==1.0.0
notebook==7.0.6
ipywidgets==8.1.1  # Interactive widgets

# -------------------- Testing (Optional) --------------------
pytest==7.4.3
pytest-cov==4.1.0

# -------------------- Type Checking (Optional) --------------------
# mypy==1.7.1
# types-PyYAML==6.0.12.12

# ============================================
# Installation Instructions:
# 
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # On Windows: venv\Scripts\activate
#
# 2. Install dependencies:
#    pip install -r requirements.txt
#
# 3. Create .env file with your Gemini API key:
#    GEMINI_API_KEY=your_api_key_here
# ============================================

